+++ fsspec/__init__.py
@@ -4,6 +4,7 @@ from .spec import AbstractFileSystem
 from .registry import get_filesystem_class, registry, filesystem
 from .mapping import FSMap, get_mapper
 from .core import open_files, get_fs_token_paths, open
+from . import caching
 
 __version__ = get_versions()["version"]
 del get_versions
@@ -19,4 +20,5 @@ __all__ = [
     "open",
     "open_files",
     "registry",
+    "caching",
 ]
